Implement basic persistent highlighting features
There are some issues with this approach. Primarily it's highlighting
based on a regex. Meaning any phrase matching will be highlighted. This
is most apparent in duplicate highlights which occur often with smaller
phrases.

I'll have to test the workflow more to see if it's worth putting more
time into this or not.